// client/chinesechess_oops/assets/script/game/enum/CHESS_TYPE.ts
export enum CHESS_TYPE {
    BLACK_C = 0x01,     //黑车
    BLACK_J = 0x02,     //黑将
    BLACK_M = 0x03,     //黑马
    BLACK_P = 0x04,     //黑炮
    BLACK_S = 0x05,     //黑士
    BLACK_X = 0x06,     //黑象
    BLACK_Z = 0x07,     //黑卒

    RED_C = 0x11,     //黑车
    RED_J = 0x12,     //黑将
    RED_M = 0x13,     //黑马
    RED_P = 0x14,     //黑炮
    RED_S = 0x15,     //黑士
    RED_X = 0x16,     //黑象
    RED_Z = 0x17,     //黑卒
}